Add loader for H5J format.
H5J is the new HDF5 based container for H.26[45] generated movies. The movies are stored 1 movie per channel. The names of the channels are not fixed and should reflect meaningful names to the data. Additional meta-data is easily added either by creating another group, or by adding additional attributes (for an attribute the data needs to < 64K). The data is stored in the HDF5 container in the following layout:
GROUP "/" {
   GROUP "Channels" {
      ATTRIBUTE "frames" {
         DATATYPE  H5T_STD_I64LE
         DATASPACE  SCALAR
      }
      ATTRIBUTE "height" {
         DATATYPE  H5T_STD_I64LE
         DATASPACE  SCALAR
      }
      ATTRIBUTE "width" {
         DATATYPE  H5T_STD_I64LE
         DATASPACE  SCALAR
      }
      DATASET "Channel_1" {
         DATATYPE  H5T_STD_U8LE
         DATASPACE  SIMPLE { ( size of channel 1 ) / ( size of channel 1 ) }
      }
      DATASET "Channel_2" {
         DATATYPE  H5T_STD_U8LE
         DATASPACE  SIMPLE { ( size of channel 2 ) / ( size of channel 2 ) }
      }
      .....
   }
}


Former-commit-id: 8abdf6d3e246fd0540295effff38c98d9e371a0a